Playerchanges

+++ java/core/src/main/java/com/angrysurfer/core/Constants.java
@@ -10,7 +10,7 @@ public class Constants {
     public static int DEFAULT_PART_LENGTH = 1;
     public static int DEFAULT_MAX_TRACKS = 16;
     public static long DEFAULT_SONG_LENGTH = Long.MAX_VALUE;
-    public static long DEFAULT_SWING = 25L;
+    public static int DEFAULT_SWING = 25;
 
     public static final int DEFAULT_LOOP_COUNT = 0;
     public static final int DEFAULT_PART_COUNT = 1;
